upgrade: Container upgrades existing data on start
Postgres will not automatically upgrade existing data when moving
to a different major version. These changes detect existing data
and compare their major versions to see if an upgrade is required.
If the upgrade is required then the correct tooling for the existing
data is installed and `pg_upgrade` is run.

Also part of this PR is the restructure of the data volume so that a
versioned sub-directory is used instead of the root of the mount. This
means that as you upgrade major versions you will keep a version of your
data which can be used.

Signed-off-by: Rich Bayliss <rich@balena.io>
Change-type: minor